Do not treat <input> tags as self-closing
This will allow <input type="text"> text </input>,
which while not valid HTML, will be used in Decode to tag
corresponding input element from the Figma design.

This is the only modification from the upstream @lezer/html repo.
There is a forked version of @codemirror/lang-html that refers to
this repo in github.com/jasim/lang-html.